requirements.txt
pytorch_lightning
timm
torchmetrics
typed-argparse
wandb
